NSIS: workaround for Windows buggy execution of uninstallers
Includes StdUtils plugin and makes use of ExecShellAsUser to restart
the uninstaller when started as administrator. This stops the
uninstaller from failing for single-user installations of standard
users, when executed elevated by 'Apps & features' of Win10. It also
gets the correct user directories when executed by a standard user from
'Add/Remove Programs' or 'Apps & features' of an all-users installation.
This has the drawback of the uninstaller having to elevate if it needs
to, even when it's started as administrator. So the user gets the UAC
prompt twice in this case.