Add workaround for accepting third-party cookies
An empty first-party URL is considered a first-party request. This workaround
is equivalent to the Qt fix: https://codereview.qt-project.org/#/c/244021/

Note that later versions of Chromium changed that check:
https://bugs.chromium.org/p/chromium/issues/detail?id=882107
That shouldn't be a problem since we disable our workaround with Qt 5.11.3, so
once Qt catches up there, the check isn't done anymore.

Fixes #4281