Move statusbar.settings validation into configtype.
The parsing and validation of the `statusbar.settings` setting is moved
into a specialised configtype. Additionally it is now a dict instead of
a list of strings that are actually pairs. The one thing we lose from
this is ordering, before you could control the order that the indicators
showed but that no longer appears to be the case.

It would be nice to be able to use the `fixed_keys` feature of the
`Dict` configtype but the parsed config object isn't available at init
time (I was surprised I could even import it without creating a loop).
We could do a `if not self.fixed_keys: self._get_bool_options()` at the
top of `to_py()` but this is only done at boot and when the option
changes (right?).

Some defensive coding was taken out of the settings statusbar widget now
that the settings framework is making sure things exist and are of the
right type.

This should be able to be entirely squashed into the first commit of
this PR and maybe the "allow querying permission states" one (don't
forget to update the commit message). I am keeping it separate for now
for ease of fiddling later.