Fix match highlighting on qt 5.11.3 and 5.12.1+
For some reason, populating the text (even if we completely overwrite
it later) before the highlighter is added results in no highlighting.
I think this has something to do with python locking/delayed
re-highlighting.

At some point, it would be nice to file this as a bug, but I can't
seem to make a small c++ example.

See https://codereview.qt-project.org/#/c/244699/4//ALL and
https://bugreports.qt.io/browse/QTBUG-71307 for more information.

Closes #4519