Properly serialize HistoryEntry.lastVisited time.
Chromium's internal time format is a int64 with the number of
microseconds since some Windows epoch 1601-01-01 00:00:00 UTC
(base/time/time.h). They warn that you shouldn't be using the internal
format directly but we don't have that luxury.

The defensive coding in miscmodels is because 1)
`QWebHistoryItem::lastVisited()` on webkit appears to always return an
invalid QDateTime 2) if the internal date is 0 for whatever reason we
will get a large negative number from `toSecsSinceEpoch()` which will
result in an OverflowError because it doesn't fit in `time_t` on 32bit
systems.

The defensive coding in tabhistory is to handle the default of
TabHistoryItem.last_visited being `None`. For instance if you boot with
the webkit backend, save (so last_visited will be '' in the session
yaml), then load the same session on webengine.