Use ordered dicts to reduce complexity of pattern setting add
Previously, adding a pattern setting was an O(n) operation, which made
large sets of pattern settings impossible.

Instead of storing pattern settings in a list, we can stored them in
an ordered map from pattern -> ScopedValue. This provides relatively
fast deletion and insertion.

With this change, the newly added benchmark of adding the existing
hosts_list measures to 1s, where it was timing out/crashing for me
when limited to 1000 entries before.

Now, the bottleneck is the initialization of UrlPattern objects.
However, if we do anything more clever for faster lookup, we will take
a penalty to init time (which is worth it).