Add cache to '_basic_str_validation'
A very expensive part of getting string config values is the string
validation. Thankfully, not many values seem to use the 'forbidden'
option.

The 'proper' fix for this would be to subclass String (and all other
immutable types), add a flag to set when fully validated, and use that
flag instead.

This is more of a band-aid until we can get a proper solution for the
config system.