Delete disconnected IPC sockets.
The docs[0] for `nextPendingConnection` say:

> The socket is created as a child of the server, which means that it is automatically deleted when the QLocalServer object is destroyed. It is still a good idea to delete the object explicitly when you are done with it, to avoid wasting memory.

And indeed it wastes a fair amount of memory. You can test it by doing,
for example,

    for _ in `seq 0 1000`;do open_url_in_instance.sh ":nop";done

For me the %MEM column in top for the qutebrowser python process rose
about 200M each time I ran that loop.

I don't think we have to worry about use after free in the C++ server
code since the docs say to do exactly this.

[0]: https://doc.qt.io/qt-5/qlocalserver.html#nextPendingConnection