Save default UA for chrome version detection.
Since we cannot create new QWebEngineProfiles willy nilly without
hitting an assert in single process mode save the default user agent
before anything has a chance to save it.

It feels like the try/except in version.py is a little broad but the
theory is that that method should only be called on the webengine
backend where all that stuff should have already been imported.

I would have made the now global all upper case but all the other ones
beside it are lower case. They are only used in that file though...