Don't init private profile in single process mode.
Constructing a new QWebEngineProfile when in single process mode will
crash the process [since[0] Qt 5.12.2. The `'--single-process' not in
args` check is already being used in
`CommandDispatcher._new_tabbed_browser()`, I didn't change that to check
instead if a private profile was actually available because that would
make it have to import backend specific stuff.

This change is inelegant and repetitive. The way to make it less so
would be to pull all the stuff that sets things on two profiles into
methods that took a profile as an argument. I'm not sure it is worth
doing that when we only support two of them however.

[0]: https://codereview.qt-project.org/#/c/248197/