Assert private profile is available in view init.
This is only because it could theoretically be `None` here if we are in
single process mode. It should never actually happen though because it
comes to the view from the tab and the tab inherits it from the
TabbedBrowser which gets in from the MainWindow. The only time a
MainWindow is created where it doesn't inherit the private flag of
another window is in the CommandDispatcher for open where we already
check if we are in single process mode.

Except for when loading a session. Then it gets the private flag from
the session file.
And of course the `content.private_browsing true` setting would go ahead
and try to use the private profile anyway.

I don't like this inelegant way of handling such errors but there is no
mechanism for reporting errors on tab creation back up the stack
gracefully currently. We just have to make sure it crashes noisely so we
can fix the cause.