Allow _chromium_version() to fall back to profile.
Allow _chromium_version() to fall back to the default profile. This to
to account for the case where `version.version()` is called from
`app.py` when the `--version` arg is passed to qutebrowser. This is
called before webenginesettings has been initialized so the profiles
havn't been setup and the default user agent hasn't been saved.
We don't do this in normal usage because it would mean we would pick up
a user-set user agent. I added tests to ensure that fallback is not used
if the backend is initialized properly.