Cache widget text per-url.
Going through an unknown amount of settings and doing an unknown amount
of work hitting the config system and feature permissions can be
expensive. This is being called often, when we switch tabs, when a URL
changes, when the load status of a page changes. But if relevant config
values haven't changed and no new permissions have been granted we don't
need to re-create the resulting text.

The default size of `lru_cache` is 128 items. We don't want it to grow
indefinitely because then it'll just grow according to your browsing
history but there is no reason for it not to be able to have an entry
for each of your tabs. So I doubled it for good luck.