Fix dotted content-disposition test on 5.13.
Download items seem to now have a more consistent extra round of percent
encoding on them. For instance with the following header

    Content-Disposition=download;%20filename%3D%2568%252F%2569

5.11: h%2Fi
5.13: h/i

I'm not sure why the %2F wasn't getting decoded previously wher the %68 and
%69 where. And what else is/was being sepecial cased. I don't see any
related changes in `src/core/download_manager_delegate_qt.cpp` (there is
a `.NormalizePathSeparatorsTo('/')` but that is supposed to be a no-op
on posix) so it must be a chromium change.

Since the path seperator is now decoded the path normalization gets rid
of it so we end up with the same result as 5.8. I added another test to
double check the doubly encoded one, I'm not sure it is necessary
though.

See https://github.com/qutebrowser/qutebrowser/issues/4592