Tabs with invalid URLs no longer get about:blank
With 5.13 the vies that eventuate from `window.open('', 'my_window')` no
longer get fire any `urlChanged` signals. With 5.11 there used to be
two. One for `about:blank` and one for `` (the latter was ignored for
tab title update purposes).

Now the tab title only tries to get set on load finished but since it is
empty it doen't result in anything.

This test was added in f2c4cedf619f8 which was to make sure invalid URLs
didn't get added to history. There are still tests that make sure that
doesn't happen this test seems to be just to ensure the tab actually
loads. It still does and the tests to ensure the URL doesn't get added
to the history still pass (and "Ignoring invalid URL being added to
history" is logged so it is not just being avoided by not firing the
expexted signals).

https://github.com/qutebrowser/qutebrowser/issues/4592