Fix tab and webkit background colors clearing on config change
Per the qt docs, we should not call setPalette on a class that also
uses stylesheets. Since #4637, we set a stylesheet on the application,
so we cannot use setPalette anymore. This commit refactors those calls
to use stylesheets instead.

As far as I can tell, QPalette is OK to use when manually drawing (we
currently use it there).