Add prototype window close undo command.
Adds a shutting_down signal to TabbedBrowser.
Application listens for that signal and hoards the closing browser's
undo stack.
A new command is added to make a new window and give it a undo stack.
Only the first tab on the undo stack is popped. We don't know how many
were open when the window was closed. Since the only way to close a
window(?) is with :close though the easiest way to get that would be to
have `TabbedBrowser.shutdown()` clear the undo stack before closing the
tabs. Then change the window undoer to undo all the tabs, or flatten the
entries into one list.

Still needs to remember the page geometry and privatness and be wired
into the existing :undo command.