Don't consider undoing into a None tab.
For undoing the closing of windows we currently instantiate a
TabbedBrowser without any tabs. We make tabs by filling in the browsers
`_undo` list and calling `undo()`. If the "tabs.last_close" setting was
on of the values that modifies the last tab instead of closing it we
were getting an AttributeError when trying to access the history of the
sole tab.