Remove empty icon workaround for QtWebEngine
This was originally introduced for QtWebKit in a loadStarted slot in
ec6066c7baf27193153f0310f88bf6af83dd80b3 (and then reintroduced in
bfcb3094607d6304c4289631d5059db904a6ce5b for #187).

In 9320214429e6dc269d7d87042e474873ab8ad4f4 it was removed for QtWebEngine as a
fix for #3469.

c9cd47b5b11bb23c043df7f997038dc7a085ee6e reintroduced it for QtWebEngine as it
still was a (small) problem for pages loading for a long time. However, it was
disabled with Qt >= 5.10 as the workaround from #3427 interferes and causes
icons to be unset after QtWebEngine set it.

Even with the workaround removed, we still see this happening - since the
issue this fixes isn't really important, and it causes more problems than it
solves, let's just get rid of it entirely.