Disable error pages in webpage fixture.
Due to a [change][] in webkit that introduced a default error page the
the webkit/test_tabhistory.py::test_titles broke with the active tab
because network access is disable in the webpage fixture.

This change disables the default error pages in the fixture by telling
the error notify thing that it doesn't support doing anything with the
errors. It also disables the `ChooseMultipleFilesExtension` which should
be used with this fixture since we override that for real in qutebrowser

I'm declaring the class inside the method because I don't know how to
skip declaring a class in stubs.py with importskip.

change: https://github.com/annulen/webkit/commit/3b3724183bbe