Add separated loader.load_extensions method.
The vulture and docs scripts both call load_components befer
standadir._init_data() is called, so they were getting a KeyError when
the extension loader went to find the extension directory. There
shouldn't be any extensions installed in those cases anyway, and if
there were they shouldn't be included in the docs.

Also the docstring on load_components was lying.