qute-pass: Add support for private domains (e.g. "myrouter.local")
Currently, qute-pass can only be used if the domain is either a public
domain (which means that its suffix needs to be included in the Public
Suffix List [PSL]), or if an IPv4 address is used.

Some uses might want to use qute-pass to login into the web interface of
their router, printer, NAS or other network-enabled device that is only
accessible from the local network. However, currently users to need
to remember or bookmark the IPv4 addresses of these devices.

If a local DNS server is used to assign domains like "mydevice.local" to
these devices, qute-pass won't work because all potential targets are
empty:

    >>> tldextract.extract('https://mydevice.local')
    ExtractResult(subdomain='mydevice', domain='local', suffix='')
    >>> tldextract.extract('https://mydevice.local').fqdn
    ''
    >>> tldextract.extract('https://mydevice.local').registered_domain
    ''
    >>> tldextract.extract('https://mydevice.local').ipv4
    ''

This adds an additional potential target by joining subdomain and domain
if (and only if) the suffix is empty.