Speculative fix for permission crash
In some scenarios (which I can't seem to reproduce), the URL we get in
shared.feature_permission is an invalid one.

In that case, we set "urlstr = None", but that means we call message.* with
url=None but option=... set. Since that's invalid usage (we can't set a
per-domain option without knowing the URL) it bails out, causing a crash.

(cherry picked from commit 068f3ded896aa70cbc644674803cdeb6389d6561)