Makefile: correct target name for manpage
There is no html file being created ever, so trying to rebuild this target
would repeatedly build it from scratch, whether it existed or was up to date or
not. Plus it simply makes no sense to ask `make` for a .html file in order to
get a .1 file.

While we are at it, add a source dependency to rebuild the manpage on updates
to the asciidoc source, thus ensuring the newly corrected target name works as
intended.