Load active tab for a session first.
This avoid the first tab being show for a little while. For some reason
this was causing the page to report as visible, probably just for a
second but long enough that the lazy session loading page detected it
and loaded the full page.

Because the first tab added to a QTabWidget will be shown because it is
the only one there this makes sure the first added one is the active
one.

Apparently we still have a line length limit of 79 chars.