Add integration test for background tab visibility.
This is just an example. Probably more stuff in place to help with
future tests like `wait_for_all_tabs_to_load()` on the `qute` fixture
would be useful.

Alternatively we could maybe switch the bdd tests to use signals like
this instead of running the browser in a separate process.

I tried to do the `run_js_async` checks in parellel by making a callback
factory that made new signals and waiting on them with `waitSignals()`
but you can't listed on unbound signals...

Things that could be done to reduce the use of objreg:
* store a list of windows in an attribute on Application
* make `_commandrunner` public on TabbedBrowser