Set webview as focus proxy of InspectorSplitter
Before this change, opening a background tab caused the focus to be set on the
InspectorSplitter rather than the tab itself.

As it makes no sense to focus the InspectorSplitter, we pass that through to
the webview, thus fixing the failing test_focused_webview_after_*_link_in_bg
tests in tabs.bdd.