fix webkit pdf.js download test
This test should not (yet) succeed on webkit, as webkit always downloads
the PDF as viewer.html, so the line "Download test.pdf" should not lead
to a successful execution of the test.

However, since the PDF has to be loaded via a download in the first
place in order to be shown, we still get a "Download test.pdf finished",
which has nothing to do with the user pressing the download button and
is merely an "implementation detail" of how qutebrowser first downloads
the PDF and then shows it in pdf.js.

In order to "fix" this, it is enough to wait for the first PDF download,
as then the final condition will ignore that line (as we've already
waited for it before). The star is needed because webengine prepends our
usual temporary-download-random-characters, and the test should work on
both backends.