pdfjs: fix filename for webkit downloads
Only looking at the URL will not produce the right result, as the pdf
will always be saved as "viewer.html", as that's the name that is in the
URL. We try to be smarter than that by checking specifically for our
pdf.js instance (qute://pdfjs/), and then extracting the correct PDF
name from there.

This means that PDFs saved from pdf.js running in webkit will now have
their proper name, and not viewer.html.