Disable logging if cmdline starts with "@".
If a commandline starts with the "@" character, all logging is disabled
during the execution of this commandline. This is useful for running
sensitive commands whose text should not appear in the logs. For
example, a password manager userscript might use:

```
echo "@insert-text $PASSWORD" >> "$QUTE_FIFO"
```

This contributes to #938, though a few issues are unadressed:

- Should this be disabled for async commands? Which commands are async?
- Does the async nature of userscripts mean this could potentially
  conceal logging for things unrelated to the command?
- Should live command completion ignore the `@`? Or should `@` be
  completely disabled for live commands, if possible? It would be a
  false sense of security for live commands, as your command will show
  up in all the completion logging as you type it. We might be able to
  disable completion logging as soon as we see a `@` though ...

Note: I initially tried adding a `:silent` command with a syntax similar
to `:later` (e.g. `:silent insert-text foo`), but the whole command line
will get logged before it is actually executed, i.e. the `silent` bit
doesn't kick in soon enough if it is a regular command.