Add a manual editor pidfile check
This also reverts 083d0af1302f87a0c41f553f89f6736f2a215a92 because it turns out
that didn't help in the end.

Not sure what's going on there. A race condition between writing the pidfile
and us waiting for the signal? An issue with the QFileSystemWatcher?

Speculatively fixes #5362.