Basic code for sending notifications over DBus on Linux.
We attempt to connect to the DBus notification daemon; if that fails, we
die catastrophically. Handling that gracefully, configuration for always
using the system tray, and so on will be in a follow-up commit.

Unfortunately we can't implement callbacks for when the notification is
clicked, due to limitations in the PyQtWebEngine wrapper.

See #4780.