asciidoc2html: Fix linking of quickstart file
The change in #5559 was wrong: We want to link quickstart.html to
doc/quickstart.html relatively *inside* the website copy, we don't want to link
to the absolute path of the git repository.

This happened to work on the old hoster, but broke after the server move to
uberspace, where the webserver process isn't allowed to access files outside of
the web document root.