Simplify private handling in MainWindow
We don't actually need the bool(private). This was added in
1c50377c0a30cc56ef723125884c15de9523044b where app._process_args() called
MainWindow(private=None), but that was actually equivalent to private=False.
I think the original intention was to use "None" to say "the setting decides",
but later the code was changed so that the setting always wins.

That private=None call was later removed in
6ca96147807dac34ada74ef0398f37ea7466bf63 and
33f2657ce0ef048f7f2011015f0db05417d138d2, thus there's no need for the
bool(...) anymore and the rest of the code can be simplified a bit.