Use URL string for parse_javascript_url
The previous approach converted "javascript:///" to "/" as Qt only sets
an internal "hasAuthority" flag, but we can't actually know whether
an (empty) authority is present or not.

However, QUrl doesn't let us get a FullyDecoded string because the value
could be ambiguous - instead, we get the FullyEncoded variant now, and
decode it via Python's urllib.parse.

This seems to fix the /// corner case, and also is much closer to the
standard.