Handle tab pinned status in AbstractTab
This replaces TabbedBrowser.set_tab_pinned() with
AbstractTab.set_pinned() and thus also allows us to set a tab as pinned
without having to know which TabbedBrowser it belongs to.

This also fixes a bug when :undo is used after closing a tab and then
setting tabs_are_windows to true - we asked "self" (i.e. the
TabbedBrowser the tab was closed in) to restore the tab's pinned status,
but the tab wasn't actually part of that TabbedBrowser as it was opened
in a new window after the undo.