[readability-js] When hinting, use the URL selected via hints
Previously, this userscript always assumed the user wanted to parse
the current page. However, this meant that something like `:hint all
userscript readability-js` wouldn't work. To fix this, the script now
checks if qutebrowser is in hinting mode, and sets the page to be
processed accordingly.

An alternative solution would be to rewrite this script to rely on
`QUTE_URL`+`JSDOM.fromURL` accross the board. However, using `QUTE_HTML`
when the target page is already loaded seems faster, uses less data, and works offline.