Re-adding history to dmenu_qutebrowser userscript
As there is not a file called history to cat from in $QUTE_DATA_DIR, but instead a sqlite file called history.sqlite.
Thus I changed this for sqlite3 and an inline query on the history table, extracting distinct urls. And with fprint and cat piping them into dmenu,
the rest works as always.