Warn when non-feature boolask settings are requested
I don't think stopping these settings from being entered is feasible
because getting the list of settings that are page features requires
having an initialised tab at this point.

AbstractPermission.test_feature() raises a WebTabError (with an
appropriate error message) when it is called but it was getting
swallowed because I was using that exception for control flow too.

_test_feature now returns None instead of raising and we flash an error
whenever the widget gets rendered and one of the BoolAsk settings it checks
isn't a page feature.