Add a setting for window transparency
On some setups, it looks like setting WA_TranslucentBackground causes
worse performance. Additionally, it also breaks dmenu window embedding
with its `-w` flag. This adds a setting to make it optional.

I originally tried to set this in a config change handler, but even with
a subsequent .hide()/.show(), for some odd reason that causes
rendering/repainting artifacts, possibly due to a Qt bug.

Follow-up to #5546
Closes #5805