Add workaround for fetching model data in completion
It looks like our views don't update correctly when fetchMore() is
called on models, see 32fa1ff and #5897.

Since Qt 5.11, Qt actually calls fetchMore in models internally when
their view is scrolled to the maximum, so that this condition doesn't
actually work properly anymore (thus regressing #2841), because
`rowCount()` on the model will return the updated row count after
already fetching more data.

As a workaround, we now ask the view whether the index below the current
one is being displayed, and if not, use the existing expandAll() hack to
force it loading new data.

This also fixes a crash introduced by the new PgUp/PgDown
bindings (`:completion-item-focus {prev,next}-page`): If we're moving
towards the end of visible data, PgDown will at some point focus the
last visible item. If we then press PgDown again, we move the selected
item to an invisible (unloaded) item, thus causing a ZeroDivisionError
when trying to get the element height of those items.

Fixes #5848