Fix content.headers.referer = 'never' with Qt 5.15
The --no-referrers flag was removed in Chromium 79:
https://chromium-review.googlesource.com/c/chromium/src/+/1808144

We now instead use the request interceptor to remove the referrer, which
is possible ever since Qt 5.12.4 and 5.13.1 and something we've wanted
to do anyways, see #4793.

The setting is still marked as "requires a restart" (and it doesn't
support URL patterns) because that's partially true: The `same-domain`
value is still handled via CLI arguments.

Fixes #5892 (together with the last commit for 'same-domain')