tests: Avoid calling webenginesettings.init()
Since #5813, we now do more work there - and there already is quite a
bit of odd patching, so this isn't really the right way forward.

For now, replicate the initialization steps we *actually* need inside
the tests. I'm not entirely happy with this solution, but it's likely we
can revisit this to find something nicer when refactoring profile
initialization in general, see #5935.

This unfortunately also means there's no good way we can do what
`test_default_user_agent_saved` did, so I ended up removing it.