Time units added to :later command
The :later command is modified to accept either an integer or time
duration with properly formatted units.
The expected format is XhYmZs, where X, Y and Z are integers
corresponding to hours, minutes or seconds respectively.
The user needs to specify at least one of the units for the command to
work.
In case of integer input, unit is assumed to be seconds.