Show set-cmd-text command bindings in completion.
Fixes #5942.

For bindings like `o -> :set-cmd-text -s :open`, the user conceptually
expects that `o` maps to `open`, yet they do not show in the command
completion UI.

With this patch, a binding to a command of form
`set-cmd-text [flags...] :<cmd> [cmdargs]` will be treated as if it were
bound to `<cmd> [cmdargs]` instead for the purpose of completion.
Bindings to `set-cmd-text --append` are ignored.