Don't crash on long search patterns.
Entering an overly long search term into the completion causes a crash:

```
PyQt5.QtCore.QRegularExpression.PatternOptions(1)) is not valid: regular expression is too large
```

This can happen, for example, when pressing `go` at the qutebrowser
error page. I'm unclear on exactly what the limit is, but 50k
characters seems to trip it on my machine. Still, it seems like any
pattern larger than a certain amount is likely an error, so I settled on
limiting it to 5k.

Fixes #5973