Show a hint widget for scroll marks.
In `jump_mark` mode, we now show a small window similar to the keyhint
window that displays the key for each local mark with its x and y
percentages. Percentages are calculated when the widget is shown, as
they can change for pages with "infinite scroll". While there's some
similarity between the Scroll and Key hint widgets, I didn't think there
was enough shared code to merit a shared base class. Global marks were
omitted under the assumption that they are less relevant, and computing
percentages is not possible if the page is no longer open. Marks are
sorted by their vertical position.

Fixes #5969.